Backend/requirements.txt
fastapi>=0.104.1
pydantic-settings>=2.1.0
ollama>=0.1.9
requests>=2.31.0
tqdm>=4.66.1
opencv-python>=4.8.1.78
PyMuPDF>=1.23.8
numpy>=1.24.3
pdf2image>=1.16.3
openai>=1.3.5
Pillow>=10.0.1
python-multipart>=0.0.6
uvicorn>=0.24.0
pytesseract>=0.3.10
langdetect>=1.0.9